# Improve perf of picking with large `pickingRadius`