--- a/test/bench/pick-layers.bench.js
+++ b/test/bench/pick-layers.bench.js
@@ -1,0 +1,74 @@
+import {Layer} from 'deck.gl/core';
+import {getClosestFromPickingBuffer} from 'deck.gl/core/lib/pick-layers';
+
+const SAMPLE_LAYERS = [new Layer()];
+const OBJECT_COLOR = [0, 10, 20, 1];
+const NULL_COLOR = [0, 0, 0, 0];
+
+const TEST_CASES = [
+  {
+    title: 'Solid',
+    data: generateSampleData({
+      pickingRadius: 10,
+      getColor: () => OBJECT_COLOR
+    })
+  },
+  {
+    title: 'Circle',
+    data: generateSampleData({
+      pickingRadius: 10,
+      getColor: (x, y) => (10 - x) * (10 - x) + (10 - y) * (10 - y) < 25 ? OBJECT_COLOR : NULL_COLOR
+    })
+  },
+  {
+    title: 'Triangle',
+    data: generateSampleData({
+      pickingRadius: 10,
+      getColor: (x, y) => x + y <= 10 ? OBJECT_COLOR : NULL_COLOR
+    })
+  },
+  {
+    title: 'Solid - Big',
+    data: generateSampleData({
+      pickingRadius: 50,
+      getColor: () => OBJECT_COLOR
+    })
+  }
+];
+
+export default function pickLayersBench(bench) {
+
+  return TEST_CASES.reduce(
+    (b, testCase) => {
+      return b.add(testCase.title, () => getClosestFromPickingBuffer(null, testCase.data));
+    },
+    bench.group('getClosestFromPickingBuffer')
+  );
+
+}
+
+function generateSampleData({pickingRadius, getColor}) {
+  const width = pickingRadius * 2;
+  const height = pickingRadius * 2;
+  const pixels = new Uint8Array(width * height * 4);
+
+  let i = 0;
+  for (let row = 0; row < height; row++) {
+    for (let col = 0; col < width; col++) {
+      const color = getColor(col, row);
+      pixels[i++] = color[0];
+      pixels[i++] = color[1];
+      pixels[i++] = color[2];
+      pixels[i++] = color[3];
+    }
+  }
+
+  return {
+    pickedColors: pixels,
+    layers: SAMPLE_LAYERS,
+    deviceX: pickingRadius,
+    deviceY: pickingRadius,
+    deviceRadius: pickingRadius,
+    deviceRect: {x: 0, y: 0, width, height}
+  };
+}