- Avoid decoding picking color and generating info objects during the traversal
- Skip entire rows

Before:

| Test | Result |
| ---- | ------ |
| Solid | 56.8k iterations/s (0.18s) |
| Circle | 95.2k iterations/s (0.10s) |
| Triangle | 147k iterations/s (0.68s) |
| Solid - Big | 5.00k iterations/s (0.20s) |

After:

| Test | Result |
| ---- | ------ |
| Solid | 826k iterations/s (0.12s) |
| Circle | 1.04M iterations/s (0.96s) |
| Triangle | 820k iterations/s (0.12s) |
| Solid - Big | 64.5k iterations/s (0.15s) |

Tested: node, browser, layer-browser